refactor(suggestions): extract components and hoist state in SuggestionsScreen
This commit refactors the suggestions feature UI to improve modularity and testability. It extracts internal composables into standalone components and updates the screen's entry point to follow state-hoisting patterns.

- **`SuggestionsScreen.kt`**:
    - Refactored `SuggestionsUiRoute` to accept `SuggestionsScreenUiState` and an event handler as parameters, removing the internal `ViewModel` and `LaunchedEffect` logic.
    - Moved the `NoProfileState` and `SuggestionsList` composables into a new `ui.components` package.
    - Simplified the file to focus on the top-level screen structure and navigation.
- **`NoProfileState.kt`**:
    - Created a new component to handle the empty state when a user profile is missing.
- **`SuggestionsList.kt`**:
    - Created a new component to render the structured `FashionAdvice` data, including summary text, outfit cards, and makeup advice cards.
